+++ terraform/prerequisites/outputs.tf
@@ -19,3 +19,23 @@ output "vault_seal_awskms_key_arn" {
 output "vault_license_secret_arn" {
   value = aws_secretsmanager_secret.vault_license.arn
 }
+
+output "vault_fqdn" {
+  value = var.vault_fqdn
+}
+
+output "route53_hosted_zone_name" {
+  value = aws_route53_zone.vault_private.name
+}
+
+output "vault_tls_cert_secret_arn" {
+  value = aws_secretsmanager_secret.vault_tls_cert.arn
+}
+
+output "vault_tls_key_secret_arn" {
+  value = aws_secretsmanager_secret.vault_tls_key.arn
+}
+
+output "vault_tls_ca_bundle_secret_arn" {
+  value = aws_secretsmanager_secret.vault_tls_ca_bundle.arn
+}